Add a very basic test.
I ended up needing to export the Traceparentheader const so that
it was accessible to this test.
